fixed a bug related to my misguided attempts to forward declare grackle types.
